refactor: move InputNames to consts
