+++ src/consts.ts
@@ -1,1 +1,14 @@
 export const backendStaticUrl = 'https://ya-praktikum.tech/api/v2/resources';
+
+export enum InputNames {
+  FIRST_NAME = 'first_name',
+  EMAIL = 'email',
+  SECOND_NAME = 'second_name',
+  DISPLAY_NAME = 'display_name',
+  PHONE = 'phone',
+  PASSWORD = 'password',
+  LOGIN = 'login',
+  OLD_PASSWORD = 'old_password',
+  NEW_PASSWORD = 'new_password',
+  REPEAT_PASSWORD = 'repeat_password',
+}
