+++ src/pages/utilities/index.ts
@@ -1,2 +1,1 @@
-export { InputNames } from './InputNames';
 export { inputValueUpdaterFactory } from './utilities';
